development back-merge 1.7.2

+++ Lib9c/Model/Stat/IBaseAndAdditionalStats.cs
@@ -4,32 +4,32 @@ namespace Nekoyume.Model.Stat
 {
     public interface IBaseAndAdditionalStats
     {
-        int BaseHP { get; }
-        int BaseATK { get; }
-        int BaseDEF { get; }
-        int BaseCRI { get; }
-        int BaseHIT { get; }
-        int BaseSPD { get; }
-        int BaseDRV { get; }
-        int BaseDRR { get; }
-        int BaseCDMG { get; }
-        int BaseArmorPenetration { get; }
-        int BaseThorn { get; }
+        long BaseHP { get; }
+        long BaseATK { get; }
+        long BaseDEF { get; }
+        long BaseCRI { get; }
+        long BaseHIT { get; }
+        long BaseSPD { get; }
+        long BaseDRV { get; }
+        long BaseDRR { get; }
+        long BaseCDMG { get; }
+        long BaseArmorPenetration { get; }
+        long BaseThorn { get; }
 
-        int AdditionalHP { get; }
-        int AdditionalATK { get; }
-        int AdditionalDEF { get; }
-        int AdditionalCRI { get; }
-        int AdditionalHIT { get; }
-        int AdditionalSPD { get; }
-        int AdditionalDRV { get; }
-        int AdditionalDRR { get; }
-        int AdditionalCDMG { get; }
-        int AdditionalArmorPenetration { get; }
-        int AdditionalThorn { get; }
+        long AdditionalHP { get; }
+        long AdditionalATK { get; }
+        long AdditionalDEF { get; }
+        long AdditionalCRI { get; }
+        long AdditionalHIT { get; }
+        long AdditionalSPD { get; }
+        long AdditionalDRV { get; }
+        long AdditionalDRR { get; }
+        long AdditionalCDMG { get; }
+        long AdditionalArmorPenetration { get; }
+        long AdditionalThorn { get; }
 
-        IEnumerable<(StatType statType, int baseValue)> GetBaseStats(bool ignoreZero = false);
-        IEnumerable<(StatType statType, int additionalValue)> GetAdditionalStats(bool ignoreZero = false);
-        IEnumerable<(StatType statType, int baseValue, int additionalValue)> GetBaseAndAdditionalStats(bool ignoreZero = false);
+        IEnumerable<(StatType statType, long baseValue)> GetBaseStats(bool ignoreZero = false);
+        IEnumerable<(StatType statType, long additionalValue)> GetAdditionalStats(bool ignoreZero = false);
+        IEnumerable<(StatType statType, long baseValue, long additionalValue)> GetBaseAndAdditionalStats(bool ignoreZero = false);
     }
 }
