+++ Lib9c/TableData/Skill/StatBuffSheet.cs
@@ -35,7 +35,7 @@ namespace Nekoyume.TableData
             public SkillTargetType TargetType { get; private set; }
             public StatType StatType { get; private set; }
             public StatModifier.OperationType OperationType { get; private set; }
-            public int Value { get; private set; }
+            public long Value { get; private set; }
             public bool IsEnhanceable { get; private set; }
 
             public override void Set(IReadOnlyList<string> fields)
@@ -49,7 +49,7 @@ namespace Nekoyume.TableData
                 // modifier
                 StatType = (StatType)Enum.Parse(typeof(StatType), fields[5]);
                 OperationType = (StatModifier.OperationType)Enum.Parse(typeof(StatModifier.OperationType), fields[6]);
-                Value = ParseInt(fields[7]);
+                Value = ParseLong(fields[7]);
 
                 if (fields.Count < 9)
                 {
